Move valid_x and allele_specific_cutoffs subcommands to pvactools command line tree
