+++ pvactools/lib/__init__.py
@@ -10,9 +10,6 @@ __all__ = [
     "download_example_data",
     "fasta_generator",
     "output_parser",
-    "valid_alleles",
-    'valid_algorithms',
-    'valid_netmhciipan_versions',
     'net_chop',
     "netmhc_stab",
     "filter",
