+++ pvactools/tools/__init__.py
@@ -1,7 +1,11 @@
 __all__ = [
+    'allele_specific_cutoffs',
+    'compare',
     'download_cwls',
     'download_wdls',
-    'compare',
+    'valid_algorithms',
+    'valid_alleles',
+    'valid_netmhciipan_versions',
 ]
 
 from . import *
