+++ pvactools/tools/main.py
@@ -11,6 +11,20 @@ def define_parser():
     subparsers = parser.add_subparsers()
 
     #add subcommands
+    allele_specific_cutoffs_parser = subparsers.add_parser(
+        "allele_specific_cutoffs",
+        help="Show the allele specific cutoffs.",
+        add_help=False,
+    )
+    allele_specific_cutoffs_parser.set_defaults(func=allele_specific_cutoffs)
+
+    compare_parser = subparsers.add_parser(
+        "compare",
+        help="Run a comparison between two output results folders",
+        add_help=False
+        )
+    compare_parser.set_defaults(func=compare)
+
     download_cwls_parser = subparsers.add_parser(
         "download_cwls",
         help="Download pVACtools CWLs for each tool's main pipeline",
@@ -25,12 +39,26 @@ def define_parser():
         )
     download_wdls_parser.set_defaults(func=download_wdls)
 
-    compare_parser = subparsers.add_parser(
-        "compare",
-        help="Run a comparison between two output results folders",
+    valid_alleles_parser = subparsers.add_parser(
+        "valid_alleles",
+        help="Show a list of valid allele names.",
         add_help=False
-        )
-    compare_parser.set_defaults(func=compare)
+    )
+    valid_alleles_parser.set_defaults(func=valid_alleles)
+
+    valid_algorithms_parser = subparsers.add_parser(
+        "valid_algorithms",
+        help="Show a list of algorithms supported given the specified species and/or allele",
+        add_help=False
+    )
+    valid_algorithms_parser.set_defaults(func=valid_algorithms)
+
+    valid_netmhciipan_versions_parser = subparsers.add_parser(
+        "valid_netmhciipan_versions",
+        help="Show a list of valid versions of NetMHCIIpan and NetMHCIIpanEL that can be used.",
+        add_help=False
+    )
+    valid_netmhciipan_versions_parser.set_defaults(func=valid_netmhciipan_versions)
 
     parser.add_argument(
         "-v", "--version",
