+++ pvactools/tools/pvacbind/__init__.py
@@ -1,14 +1,10 @@
 __all__ = [
     'run',
     'binding_filter',
-    'valid_alleles',
-    'valid_algorithms',
-    'allele_specific_cutoffs',
     'download_example_data',
     'top_score_filter',
     'net_chop',
     'netmhc_stab',
-    'valid_netmhciipan_versions',
     'calculate_reference_proteome_similarity',
     'generate_aggregated_report',
     'identify_problematic_amino_acids',
