+++ pvactools/tools/pvacfuse/__init__.py
@@ -2,15 +2,11 @@ __all__ = [
     'run',
     'binding_filter',
     'coverage_filter',
-    'valid_alleles',
-    'valid_algorithms',
-    'allele_specific_cutoffs',
     'download_example_data',
     'top_score_filter',
     'mark_genes_of_interest',
     'net_chop',
     'netmhc_stab',
-    'valid_netmhciipan_versions',
     'calculate_reference_proteome_similarity',
     'generate_protein_fasta',
     "generate_aggregated_report",
