+++ pvactools/tools/pvacfuse/main.py
@@ -78,27 +78,6 @@ def define_parser():
     )
     generate_aggregated_report_parser.set_defaults(func=generate_aggregated_report)
 
-    valid_alleles_parser = subparsers.add_parser(
-        "valid_alleles",
-        help="Show a list of valid allele names",
-        add_help=False
-    )
-    valid_alleles_parser.set_defaults(func=valid_alleles)
-
-    valid_algorithms_parser = subparsers.add_parser(
-        "valid_algorithms",
-        help="Show a list of algorithms supported given the specified species and/or allele",
-        add_help=False
-    )
-    valid_algorithms_parser.set_defaults(func=valid_algorithms)
-
-    valid_netmhciipan_versions_parser = subparsers.add_parser(
-        "valid_netmhciipan_versions",
-        help="Show a list of valid versions of NetMHCIIpan and NetMHCIIpanEL that can be used.",
-        add_help=False
-    )
-    valid_netmhciipan_versions_parser.set_defaults(func=valid_netmhciipan_versions)
-
     identify_problematic_amino_acids_parser = subparsers.add_parser(
         "identify_problematic_amino_acids",
         help="Mark problematic amino acid positions in each epitope or filter entries that have problematic amino acids.",
@@ -119,13 +98,6 @@ def define_parser():
         add_help = False
         )
     update_tiers_parser.set_defaults(func=update_tiers)
-
-    allele_specific_cutoffs_parser = subparsers.add_parser(
-        "allele_specific_cutoffs",
-        help="Show the allele specific cutoffs",
-        add_help=False,
-    )
-    allele_specific_cutoffs_parser.set_defaults(func=allele_specific_cutoffs)
 
     download_example_data_parser = subparsers.add_parser(
         "download_example_data",
