+++ pvactools/tools/pvacseq/__init__.py
@@ -1,9 +1,6 @@
 __all__ = [
     'run',
     'binding_filter',
-    'valid_alleles',
-    'valid_algorithms',
-    'allele_specific_cutoffs',
     'download_example_data',
     'coverage_filter',
     'generate_protein_fasta',
@@ -15,7 +12,6 @@ __all__ = [
     'add_ml_predictions',
     'net_chop',
     'netmhc_stab',
-    'valid_netmhciipan_versions',
     'calculate_reference_proteome_similarity',
     'transcript_filter',
     'identify_problematic_amino_acids',
