+++ pvactools/tools/pvacvector/__init__.py
@@ -1,10 +1,6 @@
 __all__ = [
     'run',
     'visualize',
-    'valid_alleles',
-    'valid_algorithms',
-    'valid_netmhciipan_versions',
-    'allele_specific_cutoffs',
     'download_example_data',
 ]
