--- tests/test_pvacbind_allele_specific_cutoffs.py
@@ -1,55 +1,0 @@
-import unittest
-import os
-import re
-import sys
-import py_compile
-from subprocess import PIPE
-from subprocess import run as subprocess_run
-
-from pvactools.tools.pvacbind import *
-from tests.utils import *
-
-def test_data_directory():
-    return os.path.join(
-        pvactools_directory(),
-        'tests',
-        'test_data',
-        'pvacbind'
-    )
-
-class PvacbindAlleleSpecificCutoffsTests(unittest.TestCase):
-    @classmethod
-    def setUpClass(cls):
-        cls.pvactools_directory = pvactools_directory()
-        cls.test_data_directory = test_data_directory()
-
-    def test_command(self):
-        pvac_script_path = os.path.join(
-            self.pvactools_directory,
-            'pvactools',
-            'tools',
-            'pvacbind',
-            "main.py"
-            )
-        usage_search = re.compile(r"usage: ")
-        result = subprocess_run([
-            sys.executable,
-            pvac_script_path,
-            'allele_specific_cutoffs',
-            '-h'
-        ], shell=False, stdout=PIPE)
-        self.assertFalse(result.returncode, "Failed `pvacbind allele_specific_cutoffs -h`")
-        self.assertRegex(result.stdout.decode(), usage_search)
-
-    def test_compiles(self):
-        compiled_run_path = py_compile.compile(os.path.join(
-            self.pvactools_directory,
-            'pvactools',
-            "tools",
-            "pvacbind",
-            "allele_specific_cutoffs.py"
-        ))
-        self.assertTrue(compiled_run_path)
-
-    def test_runs(self):
-        self.assertFalse(allele_specific_cutoffs.main([]))
